pyproject.toml
[project]
name = "runware-ui"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "David Birks", email = "david@birks.dev" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "streamlit>=1.37.1",
    "runware>=0.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/runware_ui"]
